Rename files; right-click menu behavior for reset
